pkg/apis: Add the 'omitempty' json struct tags to applicable custom resource types.
When adding the .Status field schema configuration for the PrestoTable and HiveTable CRDs, it looks like we mark some of the array types as required fields, so expressing those types as arrays and required, while the controller responsible for populated those status fields may not have any values yet. This leads to the case where we see the following error:

```bash
spec.partitions in body must be of type array: "null"
```

We can solve this in a couple of different ways:
- Add the `nullable: true` to the top-level property. This ignores the array type and those array properties.
- Preserve any unknown fields for any array types.

Both of the options described above are still not ideal, so I think adding the omitempty json struct tags is a low risk operation.